chore(subgraph): move subgraph service to dedicated folder
